For glacier region 1, change behavior to single_at_atm_topo
This isn't necessary to fix https://github.com/ESCOMP/ctsm/issues/423,
but we need to do this in order to simplify the glacier regions: We
eventually want to go back down to having just 3 glacier regions. This
change makes regions 0 and 1 identical in behavior, which will
facilitate later going down to 3 glacier regions. (I hope to see that
combining regions 0 & 1 will be bit-for-bit.)